fix: readyness check
